refactor(dto): remove SquareDto from RowDto
- size of each dto became too big
